Adds support to build Go module Docker image and update the Go stack
Go v1.4 was used and as of today Go v1.12 is the latest iteration of it.
It introduces Go modules which allows Go dev to have project outside of
the GOPATH.

Golang now has an official Docker image to help us build
containers. https://hub.docker.com/_/golang which is updated for every
iteration of the language.

It allows us now to have a much simpler Dockerfile containing only what's
required:

- use multistage build to allow container size reduction
- the compiled app is now on a `scratch` base container (no OS)
- it provides a user for security reason
- it provides ca-certs in case it's needed